refactor(deps): relocate source-introspection helpers to metadata/indexer (decouple PR-B slice 2)
Slice 2 of relocating the streamDeps helpers to internal/ (so a future
bintrail-console binary can reach them). Moves the source-introspection /
schema family out of cmd/bintrail (package main):

- validateBinlogFormat   → metadata.ValidateBinlogFormat
- validateBinlogRowImage → metadata.ValidateBinlogRowImage
- validateNoFKCascades   → metadata.ValidateNoFKCascades
- buildFKCascadeQuery     → metadata.BuildFKCascadeQuery (exported: an
  integration test inspects the raw query's flagged-schema set, which
  ValidateNoFKCascades's pass/fail return can't expose)
- ensureResolver         → metadata.EnsureResolver
- insertSchemaChange     → indexer.InsertSchemaChange (drops the duplicate
  nullOrStringVal in favor of indexer's existing private nullOrString)

Pure relocation, zero behavior change: bodies moved verbatim and exported (the
only edits are the metadata.→ self-qualifier removals inside EnsureResolver and
the nullOrStringVal→nullOrString swap). Every caller (index/agent/agent_validate/
doctor/stream's streamDeps) rewired; metadata/indexer gain NO new imports (all
deps already present). internal/streamrun untouched.

Tests follow their code: TestBuildFKCascadeQuery_* → internal/metadata;
TestNullOrStringVal_* deleted (it duplicated indexer's TestNullOrString_*); the
cmd_integration_test.go validate/ensureResolver/FK-cascade tests rewired to
metadata.*.

Verified: go build/vet (unit + integration) clean; full unit suite green; the
metadata-helper + monitor/supervisor integration tests (the latter exercises
streamDeps → streamrun.One running the relocated metadata/indexer helpers) pass
against MySQL.
