+++ internal/metadata/metadata_test.go
@@ -1,6 +1,7 @@
 package metadata
 
 import (
+	"strings"
 	"testing"
 )
 
@@ -122,11 +123,11 @@ func TestPKColumnMetas_ordering(t *testing.T) {
 // An unknown schema returns an empty (non-nil) slice.
 func TestResolverTables(t *testing.T) {
 	r := buildTestResolver(map[string]*TableMeta{
-		"appdb.users":     {Schema: "appdb", Table: "users"},
-		"appdb.orders":    {Schema: "appdb", Table: "orders"},
-		"appdb.products":  {Schema: "appdb", Table: "products"},
-		"otherdb.audits":  {Schema: "otherdb", Table: "audits"},
-		"otherdb.events":  {Schema: "otherdb", Table: "events"},
+		"appdb.users":    {Schema: "appdb", Table: "users"},
+		"appdb.orders":   {Schema: "appdb", Table: "orders"},
+		"appdb.products": {Schema: "appdb", Table: "products"},
+		"otherdb.audits": {Schema: "otherdb", Table: "audits"},
+		"otherdb.events": {Schema: "otherdb", Table: "events"},
 	})
 
 	t.Run("appdb_returns_three_tables_sorted", func(t *testing.T) {
@@ -161,4 +162,50 @@ func TestResolverTables(t *testing.T) {
 			t.Errorf("len = %d, want 0", len(got))
 		}
 	})
+}
+
+// When schemas are named explicitly, the scan is scoped to exactly those
+// schemas via a parameterized IN list — and the bintrail-internal exclusion is
+// NOT added (an explicitly named internal schema is still policed).
+func TestBuildFKCascadeQuery_withSchemas(t *testing.T) {
+	query, args := BuildFKCascadeQuery([]string{"iotcore", "billing"})
+
+	if !strings.Contains(query, "CONSTRAINT_SCHEMA IN (?,?)") {
+		t.Errorf("expected parameterized IN list, got query:\n%s", query)
+	}
+	if strings.Contains(query, "NOT IN") || strings.Contains(query, "information_schema.TABLES") {
+		t.Errorf("explicit --schemas must not add the internal-schema exclusion, got query:\n%s", query)
+	}
+	if len(args) != 2 || args[0] != "iotcore" || args[1] != "billing" {
+		t.Errorf("expected args [iotcore billing], got %v", args)
+	}
+}
+
+// With no schemas filter the scan excludes MySQL system schemas AND bintrail's
+// own index schemas. The latter are recognised structurally — a schema is
+// bintrail-internal only if it holds all of binlog_events, schema_snapshots and
+// stream_state — not by name, so an agent does not fatal-fail on bintrail's own
+// index FK cascades regardless of how the index DB is named (#347/#365).
+func TestBuildFKCascadeQuery_noSchemasExcludesInternal(t *testing.T) {
+	query, args := BuildFKCascadeQuery(nil)
+
+	if len(args) != 0 {
+		t.Errorf("expected no args for unscoped query, got %v", args)
+	}
+	for _, want := range []string{"'mysql'", "'information_schema'", "'performance_schema'", "'sys'"} {
+		if !strings.Contains(query, want) {
+			t.Errorf("expected unscoped query to exclude system schema %s, got query:\n%s", want, query)
+		}
+	}
+	// Structural detection: subquery over information_schema.TABLES requiring all
+	// three signature tables (HAVING COUNT(DISTINCT TABLE_NAME) = 3).
+	for _, want := range []string{
+		"information_schema.TABLES",
+		"'binlog_events'", "'schema_snapshots'", "'stream_state'",
+		"GROUP BY TABLE_SCHEMA HAVING COUNT(DISTINCT TABLE_NAME) = 3",
+	} {
+		if !strings.Contains(query, want) {
+			t.Errorf("expected unscoped query to contain %q, got query:\n%s", want, query)
+		}
+	}
 }
